Add modifiers support and conditional versions.
This allows adding a class to the responsive element like
`big` or `small`. `responsive-images` will set the appropriate URL
for the image/background image across different device resolutions.

There's also checks for invalid versions, which happen with
conditional versions not being generated but still providing an URL
because of the way Carrierwave works.